@@ -3,6 +3,8 @@
 #define GLFW_INCLUDE_VULKAN
 #include "VulkanContext.hpp"
 #include <GLFW/glfw3.h>
+#include <memory>
+#include "VulkanDevice.hpp"
 
 class Engine
 {
@@ -10,7 +12,8 @@ private:
     int width, height;
 
     GLFWwindow *window;
-    VulkanContext* vkContext;
+    std::shared_ptr<VulkanContext> context;
+    std::shared_ptr<VulkanDevice> device;
 
 public:
     Engine(int width, int height);
